Identify the Most Influential Question(s)

Scans all leaf nodes in the tree to find the questions that currently have the highest influence_index.

Usage

get_highest_influence(tree, top_n = 5, sort_by = "BOTH")

Arguments

tree

The main data.tree object for the analysis.

top_n

The number of top-ranked questions to return.

sort_by

A character string indicating how the prioritised questions should be sorted. Options are:

"TRUE" : Sort by the product of the node true_index for all ancestors, which measures the influence of the question if it is answered TRUE

"FALSE" : Sort by the product of the node false_index for all ancestors, which measures the influence of the question if it is answered FALSE

"BOTH" : (Default) Sort by the sum of 'TRUE' and 'FALSE' values which measures the aggregate influence of the question before the answer is known

Value

A data.frame (tibble) containing the name, question, the components of the influence index (influence_if_true, influence_if_false), and the total influence_index for the highest-influence leaf/leaves, sorted by influence.