Update a Tree Based on Answers Provided

Propagate the results up to the tree nodes based on the answers provided, and update the influence index to identify most important questions.

Usage

update_tree(tree)

Arguments

tree

The data.tree object to be modified.

Value

Returns the modified tree object invisibly, which allows for function chaining.

Examples

# Load a tree
ethical_tree <- load_tree_df(ethical)

# Internal indices before update
print(ethical_tree, "rule", "true_index", "false_index", influence = "influence_index")
#>                                    levelName rule true_index false_index
#> 1  Invest in Company X                        AND  0.2500000   1.0000000
#> 2   ¦--Financial Viability                    AND  0.5000000   1.0000000
#> 3   ¦   ¦--Profitability and Growth Signals    OR  1.0000000   0.3333333
#> 4   ¦   ¦   ¦--FIN1                                       NA          NA
#> 5   ¦   ¦   ¦--FIN2                                       NA          NA
#> 6   ¦   ¦   °--FIN3                                       NA          NA
#> 7   ¦   °--Solvency and Stability             AND  0.5000000   1.0000000
#> 8   ¦       ¦--FIN4                                       NA          NA
#> 9   ¦       °--FIN5                                       NA          NA
#> 10  ¦--Acceptable Environmental Stewardship    OR  1.0000000   0.5000000
#> 11  ¦   ¦--Has a Clean Current Record         AND  0.3333333   1.0000000
#> 12  ¦   ¦   ¦--ENV1                                       NA          NA
#> 13  ¦   ¦   ¦--ENV2                                       NA          NA
#> 14  ¦   ¦   °--ENV3                                       NA          NA
#> 15  ¦   °--Has a Credible Transition Pathway   OR  1.0000000   0.3333333
#> 16  ¦       ¦--ENV4                                       NA          NA
#> 17  ¦       ¦--ENV5                                       NA          NA
#> 18  ¦       °--ENV6                                       NA          NA
#> 19  ¦--Demonstrable Social Responsibility      OR  1.0000000   0.5000000
#> 20  ¦   ¦--Shows Excellent Internal Culture    OR  1.0000000   0.2500000
#> 21  ¦   ¦   ¦--SOC1                                       NA          NA
#> 22  ¦   ¦   ¦--SOC2                                       NA          NA
#> 23  ¦   ¦   ¦--SOC3                                       NA          NA
#> 24  ¦   ¦   °--SOC4                                       NA          NA
#> 25  ¦   °--Has a Positive External Impact     AND  0.3333333   1.0000000
#> 26  ¦       ¦--SOC5                                       NA          NA
#> 27  ¦       ¦--SOC6                                       NA          NA
#> 28  ¦       °--SOC7                                       NA          NA
#> 29  °--Strong Corporate Governance            AND  0.2000000   1.0000000
#> 30      ¦--GOV1                                           NA          NA
#> 31      ¦--GOV2                                           NA          NA
#> 32      ¦--GOV3                                           NA          NA
#> 33      ¦--GOV4                                           NA          NA
#> 34      °--GOV5                                           NA          NA
#>    influence
#> 1         NA
#> 2         NA
#> 3         NA
#> 4  0.4583333
#> 5  0.4583333
#> 6  0.4583333
#> 7         NA
#> 8  1.0625000
#> 9  1.0625000
#> 10        NA
#> 11        NA
#> 12 0.5833333
#> 13 0.5833333
#> 14 0.5833333
#> 15        NA
#> 16 0.4166667
#> 17 0.4166667
#> 18 0.4166667
#> 19        NA
#> 20        NA
#> 21 0.3750000
#> 22 0.3750000
#> 23 0.3750000
#> 24 0.3750000
#> 25        NA
#> 26 0.5833333
#> 27 0.5833333
#> 28 0.5833333
#> 29        NA
#> 30 1.0500000
#> 31 1.0500000
#> 32 1.0500000
#> 33 1.0500000
#> 34 1.0500000

ethical_tree <- update_tree(ethical_tree)

# Updated indices
print(ethical_tree, "rule", "true_index", "false_index", influence = "influence_index")
#>                                    levelName rule true_index false_index
#> 1  Invest in Company X                        AND  0.2500000   1.0000000
#> 2   ¦--Financial Viability                    AND  0.5000000   1.0000000
#> 3   ¦   ¦--Profitability and Growth Signals    OR  1.0000000   0.3333333
#> 4   ¦   ¦   ¦--FIN1                                       NA          NA
#> 5   ¦   ¦   ¦--FIN2                                       NA          NA
#> 6   ¦   ¦   °--FIN3                                       NA          NA
#> 7   ¦   °--Solvency and Stability             AND  0.5000000   1.0000000
#> 8   ¦       ¦--FIN4                                       NA          NA
#> 9   ¦       °--FIN5                                       NA          NA
#> 10  ¦--Acceptable Environmental Stewardship    OR  1.0000000   0.5000000
#> 11  ¦   ¦--Has a Clean Current Record         AND  0.3333333   1.0000000
#> 12  ¦   ¦   ¦--ENV1                                       NA          NA
#> 13  ¦   ¦   ¦--ENV2                                       NA          NA
#> 14  ¦   ¦   °--ENV3                                       NA          NA
#> 15  ¦   °--Has a Credible Transition Pathway   OR  1.0000000   0.3333333
#> 16  ¦       ¦--ENV4                                       NA          NA
#> 17  ¦       ¦--ENV5                                       NA          NA
#> 18  ¦       °--ENV6                                       NA          NA
#> 19  ¦--Demonstrable Social Responsibility      OR  1.0000000   0.5000000
#> 20  ¦   ¦--Shows Excellent Internal Culture    OR  1.0000000   0.2500000
#> 21  ¦   ¦   ¦--SOC1                                       NA          NA
#> 22  ¦   ¦   ¦--SOC2                                       NA          NA
#> 23  ¦   ¦   ¦--SOC3                                       NA          NA
#> 24  ¦   ¦   °--SOC4                                       NA          NA
#> 25  ¦   °--Has a Positive External Impact     AND  0.3333333   1.0000000
#> 26  ¦       ¦--SOC5                                       NA          NA
#> 27  ¦       ¦--SOC6                                       NA          NA
#> 28  ¦       °--SOC7                                       NA          NA
#> 29  °--Strong Corporate Governance            AND  0.2000000   1.0000000
#> 30      ¦--GOV1                                           NA          NA
#> 31      ¦--GOV2                                           NA          NA
#> 32      ¦--GOV3                                           NA          NA
#> 33      ¦--GOV4                                           NA          NA
#> 34      °--GOV5                                           NA          NA
#>    influence
#> 1         NA
#> 2         NA
#> 3         NA
#> 4  0.4583333
#> 5  0.4583333
#> 6  0.4583333
#> 7         NA
#> 8  1.0625000
#> 9  1.0625000
#> 10        NA
#> 11        NA
#> 12 0.5833333
#> 13 0.5833333
#> 14 0.5833333
#> 15        NA
#> 16 0.4166667
#> 17 0.4166667
#> 18 0.4166667
#> 19        NA
#> 20        NA
#> 21 0.3750000
#> 22 0.3750000
#> 23 0.3750000
#> 24 0.3750000
#> 25        NA
#> 26 0.5833333
#> 27 0.5833333
#> 28 0.5833333
#> 29        NA
#> 30 1.0500000
#> 31 1.0500000
#> 32 1.0500000
#> 33 1.0500000
#> 34 1.0500000

# Answer some questions
set_answer(ethical_tree, "FIN2", TRUE, 4)
#> Answer for leaf 'FIN2' set to: TRUE with confidence 4/5
set_answer(ethical_tree, "ENV2", TRUE, 3)
#> Answer for leaf 'ENV2' set to: TRUE with confidence 3/5
set_answer(ethical_tree, "SOC2", TRUE, 4)
#> Answer for leaf 'SOC2' set to: TRUE with confidence 4/5
set_answer(ethical_tree, "GOV2", FALSE, 1)
#> Answer for leaf 'GOV2' set to: FALSE with confidence 1/5

# Updated again
ethical_tree <- update_tree(ethical_tree)

# Updated indices
print(ethical_tree, "rule", "true_index", "false_index", influence = "influence_index")
#>                                    levelName rule true_index false_index
#> 1  Invest in Company X                        AND  0.5000000   1.0000000
#> 2   ¦--Financial Viability                    AND  1.0000000   1.0000000
#> 3   ¦   ¦--Profitability and Growth Signals    OR  1.0000000   0.5000000
#> 4   ¦   ¦   ¦--FIN1                                       NA          NA
#> 5   ¦   ¦   ¦--FIN2                                       NA          NA
#> 6   ¦   ¦   °--FIN3                                       NA          NA
#> 7   ¦   °--Solvency and Stability             AND  0.5000000   1.0000000
#> 8   ¦       ¦--FIN4                                       NA          NA
#> 9   ¦       °--FIN5                                       NA          NA
#> 10  ¦--Acceptable Environmental Stewardship    OR  1.0000000   0.5000000
#> 11  ¦   ¦--Has a Clean Current Record         AND  0.5000000   1.0000000
#> 12  ¦   ¦   ¦--ENV1                                       NA          NA
#> 13  ¦   ¦   ¦--ENV2                                       NA          NA
#> 14  ¦   ¦   °--ENV3                                       NA          NA
#> 15  ¦   °--Has a Credible Transition Pathway   OR  1.0000000   0.3333333
#> 16  ¦       ¦--ENV4                                       NA          NA
#> 17  ¦       ¦--ENV5                                       NA          NA
#> 18  ¦       °--ENV6                                       NA          NA
#> 19  ¦--Demonstrable Social Responsibility      OR  1.0000000   1.0000000
#> 20  ¦   ¦--Shows Excellent Internal Culture    OR  1.0000000   0.3333333
#> 21  ¦   ¦   ¦--SOC1                                       NA          NA
#> 22  ¦   ¦   ¦--SOC2                                       NA          NA
#> 23  ¦   ¦   ¦--SOC3                                       NA          NA
#> 24  ¦   ¦   °--SOC4                                       NA          NA
#> 25  ¦   °--Has a Positive External Impact     AND  0.3333333   1.0000000
#> 26  ¦       ¦--SOC5                                       NA          NA
#> 27  ¦       ¦--SOC6                                       NA          NA
#> 28  ¦       °--SOC7                                       NA          NA
#> 29  °--Strong Corporate Governance            AND  0.2500000   1.0000000
#> 30      ¦--GOV1                                           NA          NA
#> 31      ¦--GOV2                                           NA          NA
#> 32      ¦--GOV3                                           NA          NA
#> 33      ¦--GOV4                                           NA          NA
#> 34      °--GOV5                                           NA          NA
#>    influence
#> 1         NA
#> 2         NA
#> 3         NA
#> 4         NA
#> 5         NA
#> 6         NA
#> 7         NA
#> 8         NA
#> 9         NA
#> 10        NA
#> 11        NA
#> 12        NA
#> 13        NA
#> 14        NA
#> 15        NA
#> 16        NA
#> 17        NA
#> 18        NA
#> 19        NA
#> 20        NA
#> 21        NA
#> 22        NA
#> 23        NA
#> 24        NA
#> 25        NA
#> 26        NA
#> 27        NA
#> 28        NA
#> 29        NA
#> 30        NA
#> 31        NA
#> 32        NA
#> 33        NA
#> 34        NA

# Updated results
print_tree(ethical_tree)
#> Tree                                             Rule      Answer      Confidence   
#> Invest in Company X                               AND       FALSE       60% 
#> |-- Financial Viability                           AND                    
#> |   |-- Profitability and Growth Signals          OR        TRUE        90% 
#> |   |   |-- FIN1                                                         
#> |   |   |-- FIN2                                            TRUE        4 
#> |   |   `-- FIN3                                                         
#> |   `-- Solvency and Stability                    AND                    
#> |       |-- FIN4                                                         
#> |       `-- FIN5                                                         
#> |-- Acceptable Environmental Stewardship          OR                     
#> |   |-- Has a Clean Current Record                AND                    
#> |   |   |-- ENV1                                                         
#> |   |   |-- ENV2                                            TRUE        3 
#> |   |   `-- ENV3                                                         
#> |   `-- Has a Credible Transition Pathway         OR                     
#> |       |-- ENV4                                                         
#> |       |-- ENV5                                                         
#> |       `-- ENV6                                                         
#> |-- Demonstrable Social Responsibility            OR        TRUE        90% 
#> |   |-- Shows Excellent Internal Culture          OR        TRUE        90% 
#> |   |   |-- SOC1                                                         
#> |   |   |-- SOC2                                            TRUE        4 
#> |   |   |-- SOC3                                                         
#> |   |   `-- SOC4                                                         
#> |   `-- Has a Positive External Impact            AND                    
#> |       |-- SOC5                                                         
#> |       |-- SOC6                                                         
#> |       `-- SOC7                                                         
#> `-- Strong Corporate Governance                   AND       FALSE       60% 
#>     |-- GOV1                                                             
#>     |-- GOV2                                                FALSE       1 
#>     |-- GOV3                                                             
#>     |-- GOV4                                                             
#>     `-- GOV5